Implement rule about avoiding zero length allocations as IOperation analyzer and fixer as a generic syntaxgenerator-based fixer.
